+++ assets/src/Control.jsx
@@ -19,9 +19,12 @@ import { OverlayProvider } from 'react-aria'
 import { dynamicValuesAPI } from './dynamic-values'
 
 import controls from './controls-list.js'
+import DependendWrapper from './components/dependent/DependendWrapper'
+import RenderWrapper from './components/render/RenderWrapper'
 
 const Control = ({
   visibility,
+  data,
   ...props
 }) => {
 
@@ -38,7 +41,7 @@ const Control = ({
 
   const [value, setValue] = useState(props.value ?? '')
   const [isVisible, setVisibility] = useState(false) // False until we evaluate conditions
-  
+
   useEffect(() => {
     props.onChange && props.onChange(value)
   }, [value])
@@ -47,13 +50,6 @@ const Control = ({
   const ControlComponent = controls[type] ?? false
 
   if (!ControlComponent) return <></>;
-
-  const childProps = Object.assign({}, props)
-
-  delete childProps.value
-  delete childProps.onChange
-  delete childProps.class
-  delete childProps.wrapper
 
   const onChange = newValue => {
 
@@ -69,6 +65,11 @@ const Control = ({
     })
   }
   
+  /**
+   * For lisibility,it would be nice to move the visibility condition logic to a separate wrapper like 
+   * for the dependent values
+   */
+
   const evaluateVisibility = () => {
 
     if( ! visibility.condition ) {
@@ -77,7 +78,7 @@ const Control = ({
     } 
 
     // The default callback looks only for a value in other defined fields, but we can overwrite it if needed
-    const getValue = visibility.getValue ?? control.getValue
+    const getValue = data.getValue ?? control.getValue
     const result = evaluateFieldVisibility(visibility.condition, getValue)
     
     setVisibility( visibility.action !== 'hide' ? result : ! result )
@@ -96,13 +97,13 @@ const Control = ({
      */
     addEventListener('valueChange', field => {
 
-      // We rely on visibility.watcher for subfield changes (@see below)
+      // We rely on data.watcher for subfield changes (@see below)
       if( field.props?.controlType === 'subfield' ) return;
       
       // Avoid unnecessary evaluations
       if( ! triggerFields.includes(field.name) ) return;  
       
-      evaluateVisibility()
+      setTimeout(evaluateVisibility)
     })
 
     /**
@@ -110,8 +111,8 @@ const Control = ({
      *  
      * It is currently used to watch changes in subfields (repeaters, field-groups)
      */
-    if( visibility.watcher ) {
-      visibility.watcher(fieldName => {
+    if( data.watcher ) {
+      data.watcher(fieldName => {
         // Avoid unnecessary evaluations
         if( triggerFields.includes(fieldName) ) evaluateVisibility()
       })
@@ -130,15 +131,28 @@ const Control = ({
   
   return (
     <OverlayProvider { ...wrapper }>
-      <ControlComponent 
-        { ...childProps } 
-        value={ value }
-        onChange={ onChange } 
-        dynamic={ props.dynamic 
-          ? dynamicValuesAPI(value, setValue, props.dynamic) 
-          : false 
-        }
-      />
+      <RenderWrapper 
+        controlType={ props.controlType ?? 'field' }
+        name={ props.name ?? false } 
+        setValue={ setValue }
+        render={ refreshRender => (
+          <DependendWrapper 
+            refresh={ refreshRender } 
+            data={ data } 
+            controlProps={ props } 
+            renderControl={ controlProps => (
+              <ControlComponent 
+                { ...controlProps } 
+                value={ value }
+                onChange={ onChange } 
+                dynamic={ props.dynamic 
+                  ? dynamicValuesAPI(value, setValue, props.dynamic) 
+                  : false 
+                }
+              />
+            )} 
+          />
+      )} />
     </OverlayProvider>
   )
 }